-create new list tasks -rename list tasks -delete list tasks -delete each task -add new task

+++ lib/app/screen/home_screen/home_screen.dart
@@ -5,9 +5,12 @@ import 'package:intl/intl.dart';
 import 'package:todo_app/app/base/layout/base_layout.dart';
 import 'package:todo_app/app/bloc/home_screen_bloc.dart';
 import 'package:todo_app/app/bloc/state/home_screen_state.dart';
+import 'package:todo_app/app/bloc/state/task_state.dart';
+import 'package:todo_app/app/screen/all_list_screen/all_task_screen.dart';
 import 'package:todo_app/app/screen/complete_screen/complete_task_screen.dart';
 import 'package:todo_app/app/screen/incomplete_screen/incomplete_task_screen.dart';
-import 'file:///D:/FlutterProject/todo_app/lib/app/screen/all_list_screen/all_task_screen.dart';
+import 'package:todo_app/app/screen/list_tasks_screen/add_list_tasks_screen.dart';
+import 'package:todo_app/app/screen/list_tasks_screen/rename_list_tasks_screen.dart';
 import 'package:todo_app/common/common_constant.dart';
 import 'package:todo_app/common/foundation.dart';
 import 'package:todo_app/domain/entities/task_entity.dart';
@@ -22,10 +25,28 @@ class _HomeScreenState
   CupertinoTabController _controller;
   final TextEditingController _taskTitleController = TextEditingController();
   final TextEditingController _taskDetailController = TextEditingController();
+
   int get currentIndex => state?.currentIndex;
+
   DateTime get selectedDate => state?.selectedDate;
+
   bool get isExpand => state?.expandDetailField;
+
   ThemeData get themeData => Theme.of(context);
+
+  String get listName => bloc?.taskBloc?.state?.currentListTasks;
+
+  List<String> get listTasks => bloc?.taskBloc?.state?.listTasks;
+
+  List<TaskEntity> get completedTasks => bloc?.taskBloc?.state?.completedTasks;
+
+  List<TaskEntity> get uncompletedTasks =>
+      bloc?.taskBloc?.state?.uncompletedTasks;
+
+  int get totalTasks {
+    return (completedTasks?.length ?? 0) + (uncompletedTasks?.length ?? 0);
+  }
+
   @override
   HomeScreenBloc get bloc => BlocProvider.of<HomeScreenBloc>(context);
 
@@ -48,16 +69,21 @@ class _HomeScreenState
     CompletedTaskScreen(),
     UncompletedTaskScreen(),
   ];
+
   @override
   Widget buildAppBar(BuildContext context) {
     return AppBar(
-      title: Center(
-        child: Text(
-          'My Tasks',
-          style: themeData.textTheme.headline5.copyWith(color: Colors.white),
-        ),
-      ),
-    );
+        title: BlocBuilder(
+      cubit: bloc?.taskBloc,
+      builder: (context, TaskState state) {
+        return Center(
+          child: Text(
+            listName,
+            style: themeData.textTheme.headline5.copyWith(color: Colors.white),
+          ),
+        );
+      },
+    ));
   }
 
   @override
@@ -314,7 +340,8 @@ class _HomeScreenState
             child: Wrap(
               children: <Widget>[
                 _buildSectionUser(context),
-                _buildSectionTaskList(context),
+                if (listTasks.isNotEmpty == true)
+                  _buildSectionTaskList(context),
                 _buildBottomSheetItemMenu(Icons.add, 'Create new list'),
               ],
             ),
@@ -331,7 +358,14 @@ class _HomeScreenState
             child: Wrap(
               children: <Widget>[
                 InkWell(
-                  onTap: () {},
+                  onTap: () {
+                    Navigator.pushReplacement(
+                      context,
+                      MaterialPageRoute(
+                        builder: (context) => RenameListTaskScreen(),
+                      ),
+                    );
+                  },
                   child: Container(
                     width: MediaQuery.of(context).size.width,
                     padding: const EdgeInsets.fromLTRB(20.0, 10.0, 20.0, 10.0),
@@ -342,28 +376,63 @@ class _HomeScreenState
                   ),
                 ),
                 InkWell(
-                  onTap: () {},
+                  onTap: () {
+                    if (listName != defaultList) {
+                      if (completedTasks.isEmpty == true &&
+                          uncompletedTasks.isEmpty == true) {
+                        bloc?.taskBloc?.deleteList(listName: listName);
+                      }
+                      Navigator.pop(context);
+                      _showDeleteListAlert();
+                    }
+                  },
                   child: Container(
                     width: MediaQuery.of(context).size.width,
                     padding: const EdgeInsets.fromLTRB(20.0, 10.0, 20.0, 10.0),
-                    child: Text(
-                      'Delete List',
-                      style: themeData.textTheme.subtitle1,
+                    child: Column(
+                      mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+                      crossAxisAlignment: CrossAxisAlignment.start,
+                      children: [
+                        Text(
+                          'Delete List',
+                          style: themeData.textTheme.subtitle1.copyWith(
+                            color: (listName == defaultList)
+                                ? Colors.black.withOpacity(0.5)
+                                : Colors.black,
+                          ),
+                        ),
+                        if (listName == defaultList)
+                          Container(
+                            margin: const EdgeInsets.only(top: 5.0),
+                            child: Text(
+                              'The Default list can not delete',
+                              style: themeData.textTheme.subtitle1.copyWith(
+                                color: Colors.black.withOpacity(0.5),
+                              ),
+                            ),
+                          ),
+                      ],
                     ),
                   ),
                 ),
                 InkWell(
                   onTap: () {
-                    bloc?.taskBloc?.deleteAllTask(done: true);
-                    Future.delayed(const Duration(milliseconds: 1000))
-                        .whenComplete(() => Navigator.pop(context));
+                    if (completedTasks.isNotEmpty == true) {
+                      bloc?.taskBloc?.deleteAllTask(done: true);
+                      Future.delayed(const Duration(milliseconds: 1000))
+                          .whenComplete(() => Navigator.pop(context));
+                    }
                   },
                   child: Container(
                     width: MediaQuery.of(context).size.width,
                     padding: const EdgeInsets.fromLTRB(20.0, 10.0, 20.0, 20.0),
                     child: Text(
                       'Delete all complete tasks',
-                      style: themeData.textTheme.subtitle1,
+                      style: themeData.textTheme.subtitle1.copyWith(
+                        color: (completedTasks.isEmpty == true)
+                            ? Colors.black.withOpacity(0.5)
+                            : Colors.black,
+                      ),
                     ),
                   ),
                 ),
@@ -466,13 +535,21 @@ class _HomeScreenState
                         ),
                       ),
                     if (selectedDate != null)
-                      Container(
-                        decoration: BoxDecoration(
-                            borderRadius: BorderRadius.circular(10.0),
-                            border: Border.all(color: themeData.dividerColor)),
-                        padding: const EdgeInsets.fromLTRB(25.0, 0, 20.0, 8.0),
-                        child: Text(
-                          DateFormat('yMd').format(selectedDate),
+                      InkWell(
+                        onTap: () {
+                          _selectDate(context);
+                        },
+                        child: Container(
+                          margin: const EdgeInsets.only(left: 20.0),
+                          decoration: BoxDecoration(
+                              borderRadius: BorderRadius.circular(10.0),
+                              border:
+                                  Border.all(color: themeData.dividerColor)),
+                          padding:
+                              const EdgeInsets.fromLTRB(25.0, 10.0, 20.0, 10.0),
+                          child: Text(
+                            DateFormat.yMMMMd('en_US').format(selectedDate),
+                          ),
                         ),
                       ),
                     Padding(
@@ -521,7 +598,7 @@ class _HomeScreenState
                                   date: (selectedDate != null)
                                       ? DateFormat('yMd').format(selectedDate)
                                       : null,
-                                  listName: 'My Task',
+                                  listName: listName,
                                 ),
                               );
                               _taskTitleController.clear();
@@ -579,53 +656,118 @@ class _HomeScreenState
       child: ListView(
         shrinkWrap: true,
         children: [
-          _buildTaskListItem(context, 'Default'),
-          _buildTaskListItem(context, 'My Task'),
+          ...listTasks.map((e) => _buildTaskListItem(context, e)).toList(),
         ],
       ),
     );
   }
 
-  Widget _buildTaskListItem(context, String listName) {
-    return Container(
-      width: MediaQuery.of(context).size.width,
-      decoration: BoxDecoration(
-        color: Colors.blueAccent.withOpacity(0.1),
-        borderRadius: const BorderRadius.only(
-          topRight: Radius.circular(30.0),
-          bottomRight: Radius.circular(30.0),
+  Widget _buildTaskListItem(context, String title) {
+    final isActive = listName == title;
+    return InkWell(
+      onTap: () {
+        bloc?.taskBloc?.fetchData(listName: title);
+        Navigator.pop(context);
+      },
+      child: Container(
+        width: MediaQuery.of(context).size.width,
+        decoration: (isActive == true)
+            ? BoxDecoration(
+                color: Colors.blueAccent.withOpacity(0.1),
+                borderRadius: const BorderRadius.only(
+                  topRight: Radius.circular(30.0),
+                  bottomRight: Radius.circular(30.0),
+                ),
+              )
+            : const BoxDecoration(color: Colors.white),
+        padding: const EdgeInsets.fromLTRB(60.0, 20.0, 10.0, 20.0),
+        child: Text(
+          title,
+          style: themeData.textTheme.bodyText2,
         ),
-      ),
-      padding: const EdgeInsets.fromLTRB(60.0, 20.0, 10.0, 20.0),
-      child: Text(
-        listName,
-        style: themeData.textTheme.bodyText2,
       ),
     );
   }
 
   Widget _buildBottomSheetItemMenu(IconData icon, String title) {
-    return Container(
-      width: MediaQuery.of(context).size.width,
-      padding: const EdgeInsets.fromLTRB(20.0, 20.0, 20.0, 20.0),
-      decoration: BoxDecoration(
-        border: Border(
-          bottom: BorderSide(color: themeData.dividerColor),
+    return InkWell(
+      onTap: () {
+        Navigator.pushReplacement(
+          context,
+          MaterialPageRoute(
+            builder: (context) => CreateListTaskScreen(),
+          ),
+        );
+      },
+      child: Container(
+        width: MediaQuery.of(context).size.width,
+        padding: const EdgeInsets.fromLTRB(20.0, 20.0, 20.0, 20.0),
+        decoration: BoxDecoration(
+          border: Border(
+            bottom: BorderSide(color: themeData.dividerColor),
+          ),
+        ),
+        child: Row(
+          mainAxisAlignment: MainAxisAlignment.start,
+          crossAxisAlignment: CrossAxisAlignment.center,
+          children: [
+            Icon(icon),
+            const SizedBox(width: 10.0),
+            Text(
+              title,
+              style: themeData.textTheme.subtitle1,
+            ),
+          ],
         ),
       ),
-      child: Row(
-        mainAxisAlignment: MainAxisAlignment.start,
-        crossAxisAlignment: CrossAxisAlignment.center,
-        children: [
-          Icon(icon),
-          const SizedBox(width: 10.0),
-          Text(
-            title,
-            style: themeData.textTheme.subtitle1,
-          ),
-        ],
-      ),
     );
+  }
+
+  void _showDeleteListAlert() {
+    showDialog(
+        context: context,
+        builder: (context) {
+          return AlertDialog(
+            shape: const RoundedRectangleBorder(
+              borderRadius: BorderRadius.all(
+                Radius.circular(10.0),
+              ),
+            ),
+            title: const Text(
+              AlertContentString.deleteListTitle,
+            ),
+            content: Text(
+              '${AlertContentString.deleteListContent} $totalTasks '
+              '${(totalTasks == 1) ? 'task' : 'tasks'}',
+            ),
+            actions: <Widget>[
+              FlatButton(
+                onPressed: () => Navigator.pop(context),
+                child: Text(
+                  AlertContentString.cancel,
+                  style: TextStyle(
+                    color: themeData.primaryColor,
+                  ),
+                ),
+              ),
+              FlatButton(
+                onPressed: () {
+                  bloc?.taskBloc?.deleteList(listName: listName);
+                  Future.delayed(const Duration(milliseconds: 2000))
+                      .whenComplete(
+                    () => Navigator.pop(context),
+                  );
+                },
+                child: Text(
+                  AlertContentString.delete,
+                  style: TextStyle(
+                    color: themeData.primaryColor,
+                  ),
+                ),
+              ),
+            ],
+          );
+        });
   }
 }
 
@@ -633,6 +775,7 @@ class MyBottomMenuItem {
   IconData iconData;
   String title;
   bool isActive;
+
   MyBottomMenuItem({
     this.iconData,
     this.title,
